feat: zkOS Labs rebrand, honest README, Web2 wiring, execute() facade
Rebrand + runtime work (Phases 1–5 of the mainnet-readiness plan):

- Rebrand to @zkos-labs/* npm scope + zkos-labs/bastion repo URLs;
  keep bastion-agentique.fly.dev host (text rebrand only, no infra move).
- Replace README with the "Programmable Trust Runtime" vision behind
  ✅/🟡/🚧 honesty markers; preserve old README as docs/OVERVIEW.md and
  long-form vision in docs/VISION.md.
- Wire bastion-web2-firewall into the sidecar (/api/v2/evaluate-web2) with
  real enforcement of RateLimitPerProvider, ProviderBudget, CostCap and
  TimeOfDayRestriction; DID resolution no longer fabricates a stub for
  unregistered agents; surface confidential_compute on /health so the
  runtime refuses to advertise confidential while only the noop MPC client
  is active.
- Add the unified Bastion.execute() SDK facade composing sidecar policy +
  per-chain simulation + audit into pass/block/pending_hitl, with a
  confidential-privacy guard. Covered by new SDK tests.
- Capture the genuinely-absent subsystems (durable workflow engine, real
  Arcium MXE, true settlement router) as Future Epics in docs/ROADMAP.md.

Also folds in the branch's pending mainnet-readiness WIP (EVM contracts +
tests, Solana program edits, readiness docs, idl regenerations).

Verified: cargo build/test (sidecar, arcium, web2-firewall) green;
SDK tsc + jest green; mainnet audit hard gate untouched.

+++ evm/src/BastionSidecar.sol
@@ -31,12 +31,16 @@ contract BastionSidecar is IBastionSidecar, Ownable {
     /// @notice Prevents re-requesting the same transaction.
     mapping(bytes32 => bool) private _requested;
 
-    constructor(address _owner) Ownable(_owner) {
+    constructor(
+        address _owner
+    ) Ownable(_owner) {
         verifier = _owner;
     }
 
     /// @notice Set the verifier address (only owner).
-    function setVerifier(address _verifier) external onlyOwner {
+    function setVerifier(
+        address _verifier
+    ) external onlyOwner {
         verifier = _verifier;
     }
 
@@ -47,22 +51,14 @@ contract BastionSidecar is IBastionSidecar, Ownable {
     function requestEvaluate(
         address agent,
         address target,
-        uint256 value
+        uint value
     ) external returns (bytes32 requestId) {
-        requestId = keccak256(
-            abi.encodePacked(agent, target, value, block.timestamp, block.chainid)
-        );
+        requestId =
+            keccak256(abi.encodePacked(agent, target, value, block.timestamp, block.chainid));
         require(!_requested[requestId], "already requested");
         _requested[requestId] = true;
 
-        emit EvaluationRequested(
-            requestId,
-            agent,
-            target,
-            value,
-            block.chainid,
-            block.timestamp
-        );
+        emit EvaluationRequested(requestId, agent, target, value, block.chainid, block.timestamp);
     }
 
     /// @notice Fulfill an evaluation request with the sidecar's result.
@@ -90,10 +86,7 @@ contract BastionSidecar is IBastionSidecar, Ownable {
         bytes32 requestId
     ) external view returns (bool allowed, string memory reason, uint8 decision) {
         require(_requested[requestId], "not requested");
-        return (
-            evaluations[requestId],
-            evaluationReasons[requestId],
-            evaluationDecisions[requestId]
-        );
+        return
+            (evaluations[requestId], evaluationReasons[requestId], evaluationDecisions[requestId]);
     }
 }
